refactor(json): export json codec
